+++ Scripts/RunTime/TransitionSpec.cs
@@ -1,3 +1,9 @@
+/*
+ * Author: Matuyuhi
+ * Date: 2023-07-26
+ * File: TransitionSpec.cs
+ */
+
 using UnityEngine;
 
 namespace AnimationPro.RunTime