+++ Scripts/Component/SlideWithPointer.cs
@@ -1,7 +1,7 @@
 /*
  * Author: Matuyuhi
  * Date: 2023-07-26
- * File: AnimatedSlidePointer.cs
+ * File: SlideWithPointer.cs
  */
 
 using AnimationPro.RunTime;
@@ -11,7 +11,7 @@ using UnityEngine.EventSystems;
 namespace AnimationPro.Component
 {
     [RequireComponent(typeof(RectTransform)), AddComponentMenu("AnimationPro/UI/SlideWithPointer")]
-    public class AnimatedSlidePointer : SwitchableAnimationBehaviour, IPointerEnterHandler, IPointerExitHandler
+    public class SlideWithPointer : SwitchableAnimationBehaviour, IPointerEnterHandler, IPointerExitHandler
     {
         public override ContentTransform EnterTransform { get; set; }
         public override ContentTransform ExitTransform { get; set; }