+++ CHANGELOG.md
@@ -9,12 +9,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - switching state with animation
 - add name in scripts
+- add scale animation
+- popup & slide to pointer component
 ### Fixed
 - slide from API fix
 - refactor interfaces
+- refactor folder
 ### Changed
 - common parts of the AnimationBase class
-
+---
+## [1.1.0]
+### Added
+- switching state with animation
+- add name in scripts
+- add scale animation
+- popup & slide to pointer component
+### Fixed
+- slide from API fix
+- refactor interfaces
+- refactor folder
+### Changed
+- common parts of the AnimationBase class
 ---
 ## [1.0.0]
 ### Deprecated
@@ -159,7 +174,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed security issue 1
 
-[Unreleased]: https://github.com/Matuyuhi/AnimationPro/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/Matuyuhi/AnimationPro/compare/v1.1.0...HEAD
 
 [0.1.0]: https://github.com/Matuyuhi/AnimationPro/releases/tag/v0.1
 
@@ -180,3 +195,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.6.4]: https://github.com/Matuyuhi/AnimationPro/releases/tag/v0.6.4
 
 [1.0.0]: https://github.com/Matuyuhi/AnimationPro/releases/tag/v1.0.0
+
+[1.1.0]: https://github.com/Matuyuhi/AnimationPro/releases/tag/v1.1.0